Added time and threshold functionality to forecast viewer page
Accesses current time and then queries the most recent NWM forecast from that date and time.

Added function to display the selected statistical threshold on the forecast plot.  This code works but the app is currently not storing the selected stats method and it defaults to undefined when it loads the forecast page so the plot doesn't show a threshold.